# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docker Example",
        "version": "0.0.0.1"
    },
    "paths": {
        "/api/celulares": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar todos os celulares",
                "operationId": "cd45ea89fc3df08eb808a7f311828d12",
                "responses": {
                    "200": {
                        "description": "Listar todos os celulares"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Armazenar um novo celular",
                "operationId": "c5a891e2e52977a15ae31c78c7a63957",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "celulares": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "number"
                                                },
                                                "amount": {
                                                    "type": "integer"
                                                },
                                                "description": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Celulares adicionados com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/celulares/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibir um celular específico",
                "operationId": "ad9be36328376bbd9db060a7262af36a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do celular",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibir um celular específico"
                    },
                    "404": {
                        "description": "Celular não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar um celular",
                "operationId": "187ccb6a71a85caa06362a1b95cccde4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do celular",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Celular atualizado com sucesso"
                    },
                    "404": {
                        "description": "Celular não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Excluir um celular",
                "operationId": "21bf1a43ed7da0ae3f862a73c07d09ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do celular",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Celular excluído com sucesso"
                    },
                    "404": {
                        "description": "Celular não encontrado"
                    }
                }
            }
        },
        "/api/produtos": {
            "get": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Obtém a lista de todos os produtos",
                "operationId": "2c8d9ab9542b2af4e59939f07333644e",
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            }
        },
        "/api/vendas": {
            "get": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Obtém a lista de todas as vendas realizadas",
                "operationId": "f3e87d61dbe7b9924e1ec26ea1722882",
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Criação de uma nova venda",
                "operationId": "d88028b13315b5768cab14937e31874f",
                "parameters": [
                    {
                        "name": "id_venda",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON para a criação da venda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "nome": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "amount": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "A venda já existe"
                    },
                    "422": {
                        "description": "Dados não processados"
                    }
                }
            }
        },
        "/api/vendas/{id}": {
            "get": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Obtém uma venda especifica",
                "operationId": "67bc1507027596b3629d4c488dc31748",
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Edita uma venda",
                "description": "Edita uma venda existente e adiciona produtos a ela.",
                "operationId": "editarVenda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da venda a ser editada",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da venda e produtos",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "ID do produto",
                                                    "type": "integer"
                                                },
                                                "nome": {
                                                    "description": "Nome do produto",
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "description": "Preço do produto",
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "amount": {
                                                    "description": "Quantidade do produto",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Venda editada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sales_id": {
                                            "description": "ID da venda editada",
                                            "type": "integer"
                                        },
                                        "amount": {
                                            "description": "Quantidade total da venda",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "venda_id": {
                                                        "description": "ID da venda do produto",
                                                        "type": "integer"
                                                    },
                                                    "product_id": {
                                                        "description": "ID do produto",
                                                        "type": "integer"
                                                    },
                                                    "nome": {
                                                        "description": "Nome do produto",
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "description": "Preço do produto",
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "amount": {
                                                        "description": "Quantidade do produto",
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao adicionar o produto ou quantidade insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vendas"
                ],
                "summary": "Cancela uma venda especifica",
                "operationId": "620d0d0948e181653b0801cf06d7796e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da venda",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operação executada com sucesso"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            }
        }
    }
}